# resource/API/AZAZ音视频.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AZAZ音视频",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/azaz/video/upload": {
      "post": {
        "summary": "上传文件",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "上传文件(视频文件和视频封面文件都可用此接口)，返回文件url",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "a7rs283hnj4qns8y8qksv3rqc8impokq",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "example": [],
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "host": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [
                        "01HDXWP49G46X1R12NKGP2YRY1"
                      ],
                      "required": [
                        "01HDXWP49G46X1R12NKGP2YRY1"
                      ],
                      "description": "URL",
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "x-apifox-orders": [
                    "host",
                    "code",
                    "message",
                    "data"
                  ],
                  "required": [
                    "message",
                    "code"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "message": "上传成功",
                      "code": "200",
                      "host": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-119799912-run"
      }
    },
...
    "/azaz/video/showCollectList": {
      "get": {
        "summary": "收藏列表查询",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询用户的收藏列表，收藏列表只能自己查看",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "description": "当前页数，每页固定18个",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "用户id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "登录凭证",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "videoList": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/video"
                          }
                        },
                        "total": {
                          "type": "integer",
                          "title": "总记录数"
                        },
                        "currentPage": {
                          "type": "integer",
                          "title": "当前记录数"
                        }
                      },
                      "title": "通用返回对象",
                      "description": "列表可为空",
                      "x-apifox-orders": [
                        "videoList",
                        "total",
                        "currentPage"
                      ],
                      "required": [
                        "videoList",
                        "total",
                        "currentPage"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "code",
                    "message",
                    "data"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {},
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120095939-run"
      }
    },
    "/azaz/video/getPublishedVideos": {
      "get": {
        "summary": "查询发布列表",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询用户发布的视频列表",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "description": "当前页数，每页固定18个",
            "required": true,
            "example": "0",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "要查询的用户",
            "required": true,
            "example": "30",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": true,
            "example": "utvedogfgis8pm96duqgv9p8tx0b5zvw",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "videoList": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "x-apifox-refs": {
                              "01HE2DN0YDCKZVSBGXZHZNQGB7": {
                                "$ref": "#/components/schemas/video"
                              }
                            },
                            "properties": {
                              "videoId": {
                                "type": "string",
                                "title": "视频id"
                              },
                              "authorId": {
                                "type": "string",
                                "title": "作者id"
                              },
                              "coverUrl": {
                                "type": "string",
                                "title": "视频封面地址"
                              },
                              "videoUrl": {
                                "type": "string",
                                "title": "视频地址"
                              },
                              "title": {
                                "type": "string",
                                "title": "视频标题"
                              }
                            },
                            "required": [
                              "01HE2DN0YD9YWN2J89ZWMN1X0G",
                              "videoId",
                              "authorId",
                              "coverUrl",
                              "videoUrl",
                              "title"
                            ],
                            "x-apifox-orders": [
                              "01HE2DN0YDCKZVSBGXZHZNQGB7",
                              "01HE2DN0YD9YWN2J89ZWMN1X0G"
                            ],
                            "x-apifox-ignore-properties": [
                              "videoId",
                              "authorId",
                              "coverUrl",
                              "videoUrl",
                              "title"
                            ]
                          },
                          "title": "视频列表"
                        },
                        "total": {
                          "type": "integer",
                          "title": "总记录数"
                        }
                      },
                      "title": "通用返回对象",
                      "description": "列表可为空",
                      "x-apifox-orders": [
                        "videoList",
                        "total"
                      ],
                      "required": [
                        "videoList",
                        "total"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "code",
                    "message",
                    "data"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {},
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120111475-run"
      }
    },
...
    "/azaz/video/getVideos": {
      "get": {
        "summary": "视频流获取",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "获取视频流，按照时间顺序",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "lastVideoId",
            "in": "query",
            "description": "上一次最后一条视频的Id，第一次则为0",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "in": "query",
            "description": "用户点击左侧不同板块时生效，只推送对应分类，默认为0，不分类。",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "登录凭证",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "host": {
                      "type": "string",
                      "title": "主机地址",
                      "description": "一般为空",
                      "nullable": true
                    },
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "200正常，其他错误"
                    },
                    "errorMessage": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "title": "本次返回的总记录数"
                        },
                        "videoList": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "videoId": {
                                "type": "string",
                                "title": "视频id"
                              },
                              "coverUrl": {
                                "type": "string",
                                "title": "视频封面地址"
                              },
                              "videoUrl": {
                                "type": "string",
                                "title": "视频地址"
                              },
                              "title": {
                                "type": "string",
                                "title": "视频标题"
                              },
                              "likeNum": {
                                "type": "integer",
                                "title": "视频点赞数"
                              },
                              "isLiked": {
                                "type": "boolean",
                                "title": "是否点赞"
                              },
                              "commentNum": {
                                "type": "integer",
                                "title": "视频评论数"
                              },
                              "collectNum": {
                                "type": "integer",
                                "title": "视频收藏数"
                              },
                              "isCollected": {
                                "type": "boolean",
                                "title": "是否收藏"
                              },
                              "authorId": {
                                "type": "string",
                                "title": "作者id"
                              },
                              "username": {
                                "type": "string",
                                "title": "作者昵称"
                              },
                              "image": {
                                "type": "string",
                                "title": "作者头像地址"
                              },
                              "createTime": {
                                "type": "string",
                                "title": "视频发布时间",
                                "description": "yyyy-MM-DD hh/mm/ss 的格式"
                              }
                            },
                            "x-apifox-refs": {},
                            "x-apifox-orders": [
                              "videoId",
                              "coverUrl",
                              "videoUrl",
                              "title",
                              "likeNum",
                              "isLiked",
                              "commentNum",
                              "collectNum",
                              "isCollected",
                              "authorId",
                              "username",
                              "image",
                              "createTime"
                            ],
                            "required": [
                              "videoId",
                              "coverUrl",
                              "videoUrl",
                              "createTime",
                              "username",
                              "image",
                              "authorId",
                              "title",
                              "likeNum",
                              "isLiked",
                              "commentNum",
                              "collectNum",
                              "isCollected"
                            ],
                            "x-apifox-ignore-properties": []
                          },
                          "title": "视频列表",
                          "nullable": true
                        },
                        "lastVideoId": {
                          "type": "string",
                          "title": "最后一条视频的id",
                          "description": "游标，下次请求携带"
                        }
                      },
                      "title": "通用返回对象",
                      "description": "可以为空",
                      "x-apifox-orders": [
                        "total",
                        "videoList",
                        "lastVideoId"
                      ],
                      "required": [
                        "total",
                        "videoList",
                        "lastVideoId"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "host",
                    "code",
                    "errorMessage",
                    "data"
                  ],
                  "required": [
                    "code",
                    "errorMessage"
                  ],
                  "x-apifox-refs": {},
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120125290-run"
      }
    },
    "/azaz/video/publish": {
      "post": {
        "summary": "发布视频",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "发布视频，需要让用户选择对应的板块。视频封面可以选，不选默认封面",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "视频标题，不超过50个字",
            "required": true,
            "example": "IT崩盘啦，土木才是未来的铁饭碗！",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "videoUrl",
            "in": "query",
            "description": "视频地址",
            "required": true,
            "example": "http://s349n62ri.bkt.clouddn.com/f8ff07fe-5e64-4c26-9f48-be091d7bbcfd.mp4",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "coverUrl",
            "in": "query",
            "description": "视频封面地址",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "in": "query",
            "description": "视频分类，让用户选择提交到具体的板块，具体参考功能设计文档",
            "required": true,
            "example": "6",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "a7rs283hnj4qns8y8qksv3rqc8impokq",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {},
                  "x-apifox-orders": [
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "title": "通用返回对象，这里为空",
                      "description": "没有数据是为Null",
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120428841-run"
      }
    },
...
    "/azaz/video/info": {
      "get": {
        "summary": "查询视频简略信息",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询视频简略信息，用于朋友分享的展示以及点赞列表收藏列表等信息展示",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "要查询视频id",
            "required": false,
            "example": "16",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "utvedogfgis8pm96duqgv9p8tx0b5zvw",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "通用返回对象",
                      "description": "视频信息",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/video"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-orders": [
                    "01HE0F7KD678VTSHEEXYFY66DM"
                  ],
                  "x-apifox-refs": {
                    "01HE0F7KD678VTSHEEXYFY66DM": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120797292-run"
      }
    },
...
    "/azaz/video/doLike": {
      "post": {
        "summary": "视频点赞和取消点赞",
        "x-apifox-folder": "视频模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "点赞视频或者取消点赞",
        "tags": [
          "视频模块"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "点赞视频id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "1为点赞，0为取消点赞",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "authorId",
            "in": "query",
            "description": "视频作者Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "k01b6ewxj90883sh6ti4elf6nh91we9y",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    }
                  },
                  "x-apifox-orders": [
                    "data",
                    "message",
                    "code"
                  ],
                  "required": [
                    "message",
                    "code"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120827684-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "video": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "title": "视频id"
          },
          "authorId": {
            "type": "string",
            "title": "作者id"
          },
          "coverUrl": {
            "type": "string",
            "title": "视频封面地址"
          },
          "videoUrl": {
            "type": "string",
            "title": "视频地址"
          },
          "title": {
            "type": "string",
            "title": "视频标题"
          }
        },
        "title": "视频基本信息",
        "x-apifox-orders": [
          "videoId",
          "authorId",
          "coverUrl",
          "videoUrl",
          "title"
        ],
        "required": [
          "videoId",
          "coverUrl",
          "videoUrl",
          "title",
          "authorId"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
      "Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "返回码",
            "description": "2操作成功，其他错误"
          },
          "message": {
            "type": "string",
            "title": "错误信息",
            "description": "返回错误的情况向用户展示该错误信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "通用返回对象",
            "description": "没有数据是为Null",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "通用返回对象",
        "x-apifox-orders": [
          "code",
          "message",
          "data"
        ],
        "required": [
          "code",
          "message",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    }
  },
  "servers": []
}